+++ phasedemod.cpp
@@ -24,7 +24,7 @@ PhaseDemod::PhaseDemod(std::shared_ptr<SampleSource<std::complex<float>>> src) :
 
 }
 
-void PhaseDemod::work(void *input, void *output, int count, off_t sampleid)
+void PhaseDemod::work(void *input, void *output, int count, size_t sampleid)
 {
     auto in = static_cast<std::complex<float>*>(input);
     auto out = static_cast<float*>(output);
